Relocate transaction context
Transaction context is only used when building proposals so move it to a non-exported member of proposal package.
